feat(ingest/tableau): Allow specifying asset types for ingest_hidden_assets
