+++ metadata-ingestion/tests/unit/tableau/test_tableau_source.py
@@ -1,14 +1,28 @@
-from typing import Any, Dict, List
+import json
+import pathlib
+from typing import Any, Dict, List, cast
+from unittest import mock
 
 import pytest
+from freezegun import freeze_time
+from tableauserverclient import Server
 
 import datahub.ingestion.source.tableau.tableau_constant as c
+from datahub.emitter.mce_builder import DEFAULT_ENV, make_schema_field_urn
+from datahub.emitter.mcp import MetadataChangeProposalWrapper
+from datahub.ingestion.api.common import PipelineContext
+from datahub.ingestion.api.source import TestConnectionReport
 from datahub.ingestion.source.tableau.tableau import (
     DEFAULT_PAGE_SIZE,
+    SiteIdContentUrl,
+    TableauConfig,
     TableauPageSizeConfig,
     TableauSiteSource,
+    TableauSource,
+    TableauSourceReport,
 )
 from datahub.ingestion.source.tableau.tableau_common import (
+    TableauLineageOverrides,
     TableauUpstreamReference,
     get_filter_pages,
     make_filter,
@@ -16,9 +30,367 @@ from datahub.ingestion.source.tableau.tableau_common import (
     tableau_field_to_schema_field,
 )
 from datahub.metadata.com.linkedin.pegasus2avro.schema import SchemaField
+from datahub.metadata.schema_classes import (
+    DatasetLineageTypeClass,
+    FineGrainedLineageClass,
+    FineGrainedLineageDownstreamTypeClass,
+    FineGrainedLineageUpstreamTypeClass,
+    UpstreamClass,
+    UpstreamLineageClass,
+)
+from tests.test_helpers import test_connection_helpers
+
+FROZEN_TIME = "2021-12-07 07:00:00"
+
+GMS_PORT = 8080
+GMS_SERVER = f"http://localhost:{GMS_PORT}"
+
+test_resources_dir = pathlib.Path(__file__).parent
+
+config_source_default = {
+    "username": "username",
+    "password": "pass`",
+    "connect_uri": "https://do-not-connect",
+    "site": "acryl",
+    "projects": ["default", "Project 2", "Samples"],
+    "extract_project_hierarchy": False,
+    "page_size": 1000,
+    "workbook_page_size": None,
+    "ingest_tags": True,
+    "ingest_owner": True,
+    "ingest_tables_external": True,
+    "default_schema_map": {
+        "dvdrental": "public",
+        "someotherdb": "schema",
+    },
+    "platform_instance_map": {"postgres": "demo_postgres_instance"},
+    "extract_usage_stats": True,
+    "stateful_ingestion": {
+        "enabled": True,
+        "remove_stale_metadata": True,
+        "fail_safe_threshold": 100.0,
+        "state_provider": {
+            "type": "datahub",
+            "config": {"datahub_api": {"server": GMS_SERVER}},
+        },
+    },
+}
 
 
-def test_tablea_source_handles_none_nativedatatype():
+def read_response(file_name):
+    response_json_path = f"{test_resources_dir}/setup/{file_name}"
+    with open(response_json_path) as file:
+        data = json.loads(file.read())
+        return data
+
+
+@freeze_time(FROZEN_TIME)
+def test_tableau_test_connection_success():
+    with mock.patch("datahub.ingestion.source.tableau.tableau.Server"):
+        report = test_connection_helpers.run_test_connection(
+            TableauSource, config_source_default
+        )
+        test_connection_helpers.assert_basic_connectivity_success(report)
+
+
+@freeze_time(FROZEN_TIME)
+def test_tableau_test_connection_failure():
+    report = test_connection_helpers.run_test_connection(
+        TableauSource, config_source_default
+    )
+    test_connection_helpers.assert_basic_connectivity_failure(report, "Unable to login")
+
+
+def test_connection_report_test(requests_mock):
+    server_info_response = """
+        <tsResponse xmlns:t="http://tableau.com/api">
+            <t:serverInfo>
+                <t:productVersion build="build-number">foo</t:productVersion>
+                <t:restApiVersion>2.4</t:restApiVersion>
+            </t:serverInfo>
+        </tsResponse>
+
+    """
+
+    requests_mock.register_uri(
+        "GET",
+        "https://do-not-connect/api/2.4/serverInfo",
+        text=server_info_response,
+        status_code=200,
+        headers={"Content-Type": "application/xml"},
+    )
+
+    signin_response = """
+        <tsResponse xmlns:t="http://tableau.com/api">
+            <t:credentials token="fake_token">
+                <t:site id="fake_site_luid" contentUrl="fake_site_content_url"/>
+                <t:user id="fake_user_id"/>
+            </t:credentials>
+        </tsResponse>
+    """
+
+    requests_mock.register_uri(
+        "POST",
+        "https://do-not-connect/api/2.4/auth/signin",
+        text=signin_response,
+        status_code=200,
+        headers={"Content-Type": "application/xml"},
+    )
+
+    user_by_id_response = """
+        <tsResponse xmlns:t="http://tableau.com/api">
+          <t:user id="user-id" name="foo@abc.com" siteRole="SiteAdministratorExplorer" />
+        </tsResponse>
+    """
+
+    requests_mock.register_uri(
+        "GET",
+        "https://do-not-connect/api/2.4/sites/fake_site_luid/users/fake_user_id",
+        text=user_by_id_response,
+        status_code=200,
+        headers={"Content-Type": "application/xml"},
+    )
+
+    report: TestConnectionReport = TableauSource.test_connection(config_source_default)
+
+    assert report
+    assert report.capability_report
+    assert report.capability_report.get(c.SITE_PERMISSION)
+    assert report.capability_report[c.SITE_PERMISSION].capable
+
+    # Role other than SiteAdministratorExplorer
+    user_by_id_response = """
+        <tsResponse xmlns:t="http://tableau.com/api">
+          <t:user id="user-id" name="foo@abc.com" siteRole="Explorer" />
+        </tsResponse>
+    """
+
+    requests_mock.register_uri(
+        "GET",
+        "https://do-not-connect/api/2.4/sites/fake_site_luid/users/fake_user_id",
+        text=user_by_id_response,
+        status_code=200,
+        headers={"Content-Type": "application/xml"},
+    )
+
+    report = TableauSource.test_connection(config_source_default)
+
+    assert report
+    assert report.capability_report
+    assert report.capability_report.get(c.SITE_PERMISSION)
+    assert report.capability_report[c.SITE_PERMISSION].capable is False
+    assert (
+        report.capability_report[c.SITE_PERMISSION].failure_reason
+        == "The user does not have the `Site Administrator Explorer` role. Their current role is Explorer."
+    )
+
+
+def test_tableau_no_verify():
+    # This test ensures that we can connect to a self-signed certificate
+    # when ssl_verify is set to False.
+
+    source = TableauSource.create(
+        {
+            "connect_uri": "https://self-signed.badssl.com/",
+            "ssl_verify": False,
+            "site": "bogus",
+            # Credentials
+            "username": "bogus",
+            "password": "bogus",
+        },
+        PipelineContext(run_id="0"),
+    )
+    list(source.get_workunits())
+
+    report = source.get_report().as_string()
+    assert "SSL" not in report
+    assert "Unable to login" in report
+
+
+def test_tableau_unsupported_csql():
+    context = PipelineContext(run_id="0", pipeline_name="test_tableau")
+    config_dict = config_source_default.copy()
+    del config_dict["stateful_ingestion"]
+    config = TableauConfig.parse_obj(config_dict)
+    config.extract_lineage_from_unsupported_custom_sql_queries = True
+    config.lineage_overrides = TableauLineageOverrides(
+        database_override_map={"production database": "prod"}
+    )
+
+    def check_lineage_metadata(
+        lineage, expected_entity_urn, expected_upstream_table, expected_cll
+    ):
+        mcp = cast(MetadataChangeProposalWrapper, list(lineage)[0].metadata)
+
+        expected = UpstreamLineageClass(
+            upstreams=[
+                UpstreamClass(
+                    dataset=expected_upstream_table,
+                    type=DatasetLineageTypeClass.TRANSFORMED,
+                )
+            ],
+            fineGrainedLineages=[
+                FineGrainedLineageClass(
+                    upstreamType=FineGrainedLineageUpstreamTypeClass.FIELD_SET,
+                    upstreams=[
+                        make_schema_field_urn(expected_upstream_table, upstream_column)
+                    ],
+                    downstreamType=FineGrainedLineageDownstreamTypeClass.FIELD,
+                    downstreams=[
+                        make_schema_field_urn(expected_entity_urn, downstream_column)
+                    ],
+                )
+                for upstream_column, downstream_column in expected_cll.items()
+            ],
+        )
+        assert mcp.entityUrn == expected_entity_urn
+
+        actual_aspect = mcp.aspect
+        assert actual_aspect == expected
+
+    csql_urn = "urn:li:dataset:(urn:li:dataPlatform:tableau,09988088-05ad-173c-a2f1-f33ba3a13d1a,PROD)"
+    expected_upstream_table = "urn:li:dataset:(urn:li:dataPlatform:bigquery,my_bigquery_project.invent_dw.UserDetail,PROD)"
+    expected_cll = {
+        "user_id": "user_id",
+        "source": "source",
+        "user_source": "user_source",
+    }
+
+    site_source = TableauSiteSource(
+        config=config,
+        ctx=context,
+        platform="tableau",
+        site=SiteIdContentUrl(site_id="id1", site_content_url="site1"),
+        report=TableauSourceReport(),
+        server=Server("https://test-tableau-server.com"),
+    )
+
+    lineage = site_source._create_lineage_from_unsupported_csql(
+        csql_urn=csql_urn,
+        csql={
+            "query": "SELECT user_id, source, user_source FROM (SELECT *, ROW_NUMBER() OVER (partition BY user_id ORDER BY __partition_day DESC) AS rank_ FROM invent_dw.UserDetail ) source_user WHERE rank_ = 1",
+            "isUnsupportedCustomSql": "true",
+            "connectionType": "bigquery",
+            "database": {
+                "name": "my_bigquery_project",
+                "connectionType": "bigquery",
+            },
+        },
+        out_columns=[],
+    )
+    check_lineage_metadata(
+        lineage=lineage,
+        expected_entity_urn=csql_urn,
+        expected_upstream_table=expected_upstream_table,
+        expected_cll=expected_cll,
+    )
+
+    # With database as None
+    lineage = site_source._create_lineage_from_unsupported_csql(
+        csql_urn=csql_urn,
+        csql={
+            "query": "SELECT user_id, source, user_source FROM (SELECT *, ROW_NUMBER() OVER (partition BY user_id ORDER BY __partition_day DESC) AS rank_ FROM my_bigquery_project.invent_dw.UserDetail ) source_user WHERE rank_ = 1",
+            "isUnsupportedCustomSql": "true",
+            "connectionType": "bigquery",
+            "database": None,
+        },
+        out_columns=[],
+    )
+    check_lineage_metadata(
+        lineage=lineage,
+        expected_entity_urn=csql_urn,
+        expected_upstream_table=expected_upstream_table,
+        expected_cll=expected_cll,
+    )
+
+
+def test_lineage_overrides():
+    # Simple - specify platform instance to presto table
+    assert (
+        TableauUpstreamReference(
+            "presto_catalog",
+            "test-database-id",
+            "test-schema",
+            "test-table",
+            "presto",
+        ).make_dataset_urn(
+            env=DEFAULT_ENV, platform_instance_map={"presto": "my_presto_instance"}
+        )
+        == "urn:li:dataset:(urn:li:dataPlatform:presto,my_presto_instance.presto_catalog.test-schema.test-table,PROD)"
+    )
+
+    # Transform presto urn to hive urn
+    # resulting platform instance for hive = mapped platform instance + presto_catalog
+    assert (
+        TableauUpstreamReference(
+            "presto_catalog",
+            "test-database-id",
+            "test-schema",
+            "test-table",
+            "presto",
+        ).make_dataset_urn(
+            env=DEFAULT_ENV,
+            platform_instance_map={"presto": "my_instance"},
+            lineage_overrides=TableauLineageOverrides(
+                platform_override_map={"presto": "hive"},
+            ),
+        )
+        == "urn:li:dataset:(urn:li:dataPlatform:hive,my_instance.presto_catalog.test-schema.test-table,PROD)"
+    )
+
+    # transform hive urn to presto urn
+    assert (
+        TableauUpstreamReference(
+            None,
+            None,
+            "test-schema",
+            "test-table",
+            "hive",
+        ).make_dataset_urn(
+            env=DEFAULT_ENV,
+            platform_instance_map={"hive": "my_presto_instance.presto_catalog"},
+            lineage_overrides=TableauLineageOverrides(
+                platform_override_map={"hive": "presto"},
+            ),
+        )
+        == "urn:li:dataset:(urn:li:dataPlatform:presto,my_presto_instance.presto_catalog.test-schema.test-table,PROD)"
+    )
+
+
+def test_database_hostname_to_platform_instance_map():
+    # Simple - snowflake table
+    assert (
+        TableauUpstreamReference(
+            "test-database-name",
+            "test-database-id",
+            "test-schema",
+            "test-table",
+            "snowflake",
+        ).make_dataset_urn(env=DEFAULT_ENV, platform_instance_map={})
+        == "urn:li:dataset:(urn:li:dataPlatform:snowflake,test-database-name.test-schema.test-table,PROD)"
+    )
+
+    # Finding platform instance based off hostname to platform instance mappings
+    assert (
+        TableauUpstreamReference(
+            "test-database-name",
+            "test-database-id",
+            "test-schema",
+            "test-table",
+            "snowflake",
+        ).make_dataset_urn(
+            env=DEFAULT_ENV,
+            platform_instance_map={},
+            database_hostname_to_platform_instance_map={
+                "test-hostname": "test-platform-instance"
+            },
+            database_server_hostname_map={"test-database-id": "test-hostname"},
+        )
+        == "urn:li:dataset:(urn:li:dataPlatform:snowflake,test-platform-instance.test-database-name.test-schema.test-table,PROD)"
+    )
+
+
+def test_tableau_source_handles_none_nativedatatype():
     field: Dict[str, Any] = {
         "__typename": "CalculatedField",
         "id": "abcd",
